+++ src/main/java/net/notcoded/wayfix/mixin/WindowMixin.java
@@ -1,13 +1,18 @@
 package net.notcoded.wayfix.mixin;
 
+import net.minecraft.client.WindowEventHandler;
+import net.minecraft.client.WindowSettings;
 import net.minecraft.client.util.Monitor;
 import net.minecraft.client.util.MonitorTracker;
 import net.minecraft.client.util.Window;
 import net.notcoded.wayfix.WayFix;
 import net.notcoded.wayfix.config.ModConfig;
 import net.notcoded.wayfix.util.DesktopFileInjector;
+import net.notcoded.wayfix.util.WindowHelper;
 import org.lwjgl.glfw.GLFW;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -37,25 +42,37 @@ import net.minecraft.resource.InputSupplier;
 */
 
 @Mixin(Window.class)
-public class WindowMixin {
+public abstract class WindowMixin {
+    @Shadow protected abstract void onWindowPosChanged(long window, int x, int y);
+
+    @Shadow @Final private long handle;
+
+    // forge only allows injecting into constructors on return, this is a jank fix, but it works
+    //? if forge {
+    /*@Redirect(method = "<init>", at = @At(value = "INVOKE", target = "Lorg/lwjgl/glfw/GLFW;glfwDefaultWindowHints()V", remap = false))
+    *///?} else {
     @Inject(method = "<init>", at = @At(value = "INVOKE", target = "Lorg/lwjgl/glfw/GLFW;glfwDefaultWindowHints()V", shift = At.Shift.AFTER, remap = false))
     private void onWindowHints(CallbackInfo ci) {
+    //?}
+    //? if forge {
+    /*private void onWindowHints() {
+        GLFW.glfwDefaultWindowHints();
+    *///?}
         if (isWayland()) {
             GLFW.glfwWindowHint(GLFW.GLFW_FOCUS_ON_SHOW, GLFW.GLFW_FALSE);
 
-            if(!WayFix.config.injectIcon) return; // assuming that user wants wayland icon and not broken icon (mc launcher)
-            DesktopFileInjector.inject();
+            if(WayFix.config.injectIcon) DesktopFileInjector.inject();
             GLFW.glfwWindowHintString(GLFW.GLFW_WAYLAND_APP_ID, DesktopFileInjector.APP_ID);
         }
     }
 
     @Redirect(method = "updateWindowRegion", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/util/MonitorTracker;getMonitor(Lnet/minecraft/client/util/Window;)Lnet/minecraft/client/util/Monitor;"))
     private Monitor fixWrongMonitor(MonitorTracker instance, Window window) {
-        return WayFix.config.fullscreen.useMonitorName ? wayFix$getMonitor(instance) : instance.getMonitor(window);
+        return WayFix.config.fullscreen.useMonitorName && !WindowHelper.canUseWindowHelper ? wayfix$getMonitor(instance) : instance.getMonitor(window);
     }
 
     @Unique
-    private Monitor wayFix$getMonitor(MonitorTracker instance) {
+    private Monitor wayfix$getMonitor(MonitorTracker instance) {
         String monitorName = WayFix.config.fullscreen.monitorName;
         long monitorID = monitorName.trim().isEmpty() ? GLFW.glfwGetPrimaryMonitor() : ModConfig.Monitors.getMonitor(monitorName);
 
@@ -66,6 +83,18 @@ public class WindowMixin {
         }
 
         return instance.getMonitor(monitorID);
+    }
+
+
+    // KDE Plasma ONLY
+    @Inject(method = "updateWindowRegion", at = @At("HEAD"))
+    private void fixWrongMonitor(CallbackInfo ci) {
+        if(!WindowHelper.canUseWindowHelper) return;
+
+        int[] pos = WindowHelper.getWindowPos();
+        if(pos == null) return;
+
+        onWindowPosChanged(this.handle, pos[0], pos[1]);
     }
     
     @Inject(method = "setIcon", at = @At("HEAD"), cancellable = true)