+++ src/main/java/net/notcoded/wayfix/platforms/fabric/ModMenuIntegration.java
@@ -1,9 +1,10 @@
-package net.notcoded.wayfix.config;
-
 //? if fabric {
-/*import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+/*package net.notcoded.wayfix.platforms.fabric;
+
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
 import me.shedaniel.autoconfig.AutoConfig;
+import net.notcoded.wayfix.config.ModConfig;
 
 public class ModMenuIntegration implements ModMenuApi {
     @Override