+++ src/main/java/net/notcoded/wayfix/platforms/fabric/WayFixFabric.java
@@ -1,0 +1,13 @@
+//? if fabric {
+/*package net.notcoded.wayfix.platforms.fabric;
+
+import net.fabricmc.api.ClientModInitializer;
+import net.notcoded.wayfix.WayFix;
+
+public class WayFixFabric implements ClientModInitializer {
+	@Override
+	public void onInitializeClient() {
+		WayFix.init();
+	}
+}
+*///?}