+++ src/main/java/net/notcoded/wayfix/platforms/forge/WayFixForge.java
@@ -1,0 +1,38 @@
+//? if forge {
+/*package net.notcoded.wayfix.platforms.forge;
+
+import me.shedaniel.autoconfig.AutoConfig;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+import net.minecraftforge.fml.DistExecutor;
+import net.minecraftforge.fml.ModLoadingContext;
+import net.minecraftforge.fml.common.Mod;
+import net.notcoded.wayfix.config.ModConfig;
+
+//? if >=1.19 {
+import net.minecraftforge.client.ConfigScreenHandler;
+//?} elif <1.19 {
+/^import net.minecraftforge.fml.ExtensionPoint;
+ ^///?}
+
+@Mod("wayfix")
+@OnlyIn(Dist.CLIENT)
+public class WayFixForge {
+	public WayFixForge() {
+        DistExecutor.safeRunWhenOn(Dist.CLIENT, () -> WayFixForge::setupConfigScreen);
+	}
+
+    private static void setupConfigScreen() {
+        //? if >=1.19 {
+        ModLoadingContext.get().registerExtensionPoint(ConfigScreenHandler.ConfigScreenFactory.class, () ->
+                new ConfigScreenHandler.ConfigScreenFactory(
+                        (client, parent) -> AutoConfig.getConfigScreen(ModConfig.class, parent).get()
+                )
+        );
+        //?} elif <1.19 {
+        /^ModLoadingContext.get().registerExtensionPoint(ExtensionPoint.CONFIGGUIFACTORY, () ->
+                (client, parent) -> AutoConfig.getConfigScreen(ModConfig.class, parent).get()
+        );^///?}
+    }
+}
+*///?}