+++ CHANGES.md
@@ -1,2 +1,4 @@
 changes:
-- fix auto gui scaling looking broken with weird window sizes
+- <ins>**forge and neoforge support!!!!!**</ins>
+- fix key combinations not being detected while caps lock is on
+  - [Pull Request by d1BG](https://github.com/not-coded/WayFix/pull/24)